+++ src/EnquiryPageController.php
@@ -1,12 +1,11 @@
 <?php
 namespace Axllent\EnquiryPage;
 
-use PageController;
 use Axllent\EnquiryPage\Forms\CaptchaField;
+use PageController;
 use SilverStripe\Control\Director;
 use SilverStripe\Control\Email\Email;
 use SilverStripe\Control\HTTPResponse;
-use SilverStripe\Control\Session;
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Forms\CheckboxField;
 use SilverStripe\Forms\CheckboxSetField;
@@ -15,10 +14,6 @@ use SilverStripe\Forms\EmailField;
 use SilverStripe\Forms\FieldList;
 use SilverStripe\Forms\Form;
 use SilverStripe\Forms\FormAction;
-use SilverStripe\Forms\GridField\GridField;
-use SilverStripe\Forms\GridField\GridFieldConfig_RecordEditor;
-use SilverStripe\Forms\HeaderField;
-use SilverStripe\Forms\HTMLEditor\HTMLEditorField;
 use SilverStripe\Forms\HTMLReadonlyField;
 use SilverStripe\Forms\LiteralField;
 use SilverStripe\Forms\OptionSetField;
@@ -31,27 +26,44 @@ use SilverStripe\View\Requirements;
 
 class EnquiryPageController extends PageController
 {
+    /**
+     * Allowed actions
+     *
+     * @var    array
+     * @config
+     */
     private static $allowed_actions = [
         'EnquiryForm',
-        'captcha'
+        'captcha',
     ];
 
+    /**
+     * Index
+     *
+     * @param HTTPRequest $request HTTP request
+     *
+     * @return HTTPResponse
+     */
     public function index($request)
     {
         return $this->renderWith(['EnquiryPage', 'Page']);
     }
 
-    public function EnquiryForm()
+    /**
+     * Enquiry form
+     *
+     * @return Form
+     */
+    public function enquiryForm()
     {
-        if (
-            !Email::is_valid_address($this->EmailTo) ||
-            !Email::is_valid_address($this->EmailFrom)
+        if (!Email::is_valid_address($this->EmailTo)
+            || !Email::is_valid_address($this->EmailFrom)
         ) {
             return false;
         }
 
         if (!$this->EmailSubject) {
-            $this->EmailSubject = 'Website Enquiry';
+            $this->EmailSubject = 'Website enquiry';
         }
 
         $elements = $this->EnquiryFormFields();
@@ -62,17 +74,17 @@ class EnquiryPageController extends PageController
         }
 
         /* Build the fieldlist */
-        $fields = FieldList::create();
-        $validator = RequiredFields::create();
+        $fields      = FieldList::create();
+        $validator   = RequiredFields::create();
         $jsValidator = [];
 
         /* Create filter for possible $_GET parameters / pre-population */
         $get_param_filter = URLSegmentFilter::create();
 
         foreach ($elements as $el) {
-            $key = $this->keyGen($el->FieldName, $el->SortOrder);
+            $key   = $this->keyGen($el->FieldName, $el->SortOrder);
             $field = false;
-            $type = false;
+            $type  = false;
             if ($el->FieldType == 'Text') {
                 if ($el->FieldOptions == 1) {
                     $field = TextField::create($key, $el->FieldName);
@@ -83,7 +95,12 @@ class EnquiryPageController extends PageController
             } elseif ($el->FieldType == 'Email') {
                 $field = EmailField::create($key, $el->FieldName);
             } elseif ($el->FieldType == 'Select') {
-                $options = preg_split('/\n\r?/', $el->FieldOptions, -1, PREG_SPLIT_NO_EMPTY);
+                $options = preg_split(
+                    '/\n\r?/',
+                    $el->FieldOptions,
+                    -1,
+                    PREG_SPLIT_NO_EMPTY
+                );
                 if (count($options) > 0) {
                     $tmp = [];
                     foreach ($options as $o) {
@@ -93,7 +110,12 @@ class EnquiryPageController extends PageController
                     $field->setEmptyString('[ Please Select ]');
                 }
             } elseif ($el->FieldType == 'Checkbox') {
-                $options = preg_split('/\n\r?/', $el->FieldOptions, -1, PREG_SPLIT_NO_EMPTY);
+                $options = preg_split(
+                    '/\n\r?/',
+                    $el->FieldOptions,
+                    -1,
+                    PREG_SPLIT_NO_EMPTY
+                );
                 if (count($options) == 1) {
                     $field = CheckboxField::create($key, trim(reset($options)));
                 } elseif (count($options) > 0) {
@@ -104,7 +126,12 @@ class EnquiryPageController extends PageController
                     $field = CheckboxSetField::create($key, $el->FieldName, $tmp);
                 }
             } elseif ($el->FieldType == 'Radio') {
-                $options = preg_split('/\n\r?/', $el->FieldOptions, -1, PREG_SPLIT_NO_EMPTY);
+                $options = preg_split(
+                    '/\n\r?/',
+                    $el->FieldOptions,
+                    -1,
+                    PREG_SPLIT_NO_EMPTY
+                );
                 if (count($options) > 0) {
                     $tmp = [];
                     foreach ($options as $o) {
@@ -112,29 +139,30 @@ class EnquiryPageController extends PageController
                     }
                     $field = OptionsetField::create($key, $el->FieldName, $tmp);
                 }
-            } elseif ($el->FieldType == 'Header') { // Readonly field
-                $html = ShortcodeParser::get_active()->parse($el->FieldOptions);
+            } elseif ($el->FieldType == 'Header') {
+                // Readonly field
+                $html  = ShortcodeParser::get_active()->parse($el->FieldOptions);
                 $field = HTMLReadonlyField::create($key, $el->FieldName, $html);
-            } elseif ($el->FieldType == 'HTML') { // backwards compatible for old values
-                $html = ShortcodeParser::get_active()->parse($el->FieldOptions);
+            } elseif ($el->FieldType == 'HTML') {
+                // backwards compatible for old values
+                $html  = ShortcodeParser::get_active()->parse($el->FieldOptions);
                 $field = LiteralField::create($key, $html);
             }
 
             if ($field) {
-                /* Allow $_GET parameters to pre-populate fields */
+                // Allow $_GET parameters to pre-populate fields
                 $request = $this->request;
                 $get_var = $get_param_filter->filter($el->FieldName);
-                if (
-                    !$request->isPOST() &&
-                    !$field->Value() &&
-                    null != $request->getVar($get_var)
+                if (!$request->isPOST()
+                    && !$field->Value()
+                    && null != $request->getVar($get_var)
                 ) {
                     $field->setValue($request->getVar($get_var));
                 }
 
                 if ($el->RequiredField == 1) {
                     $field->addExtraClass('required');
-                    /* Add "Required" next to field" */
+                    // add "Required" next to field
                     $validator->addRequiredField($key);
                     $jsValidator[$key] = $el->FieldType;
                 }
@@ -155,7 +183,11 @@ class EnquiryPageController extends PageController
             $jsValidator['CaptchaImage'] = 'Text';
 
             if ($this->CaptchaHelp) {
-                $field->setRightTitle('<span id="CaptchaHelp">'.htmlspecialchars($this->CaptchaHelp).'</span>');
+                $field->setRightTitle(
+                    '<span id="CaptchaHelp">' .
+                    htmlspecialchars($this->CaptchaHelp) .
+                    '</span>'
+                );
             }
 
             $fields->push($field);
@@ -165,23 +197,41 @@ class EnquiryPageController extends PageController
             FormAction::create('SendEnquiryForm', $this->EmailSubmitButtonText)
         );
 
-        if (Config::inst()->get('Axllent\EnquiryPage\EnquiryPage', 'js_validation')) {
-            Requirements::customScript('var EnquiryFormValidator='.json_encode($jsValidator).';');
-            Requirements::javascript('axllent/silverstripe-enquiry-page: javascript/enquiryform.js');
+        if (Config::inst()->get(
+            'Axllent\EnquiryPage\EnquiryPage',
+            'js_validation'
+        )
+        ) {
+            Requirements::customScript(
+                'var EnquiryFormValidator=' . json_encode($jsValidator) . ';'
+            );
+            Requirements::javascript(
+                'axllent/silverstripe-enquiry-page: javascript/enquiryform.js'
+            );
         }
 
         $form = Form::create($this, 'EnquiryForm', $fields, $actions, $validator);
+
         return $form;
     }
 
-
-    public function SendEnquiryForm($data, $form)
+    /**
+     * Send enquiry form
+     *
+     * @param array $data Form data
+     * @param Form  $form Form
+     *
+     * @return HTTPResponse
+     */
+    public function sendEnquiryForm($data, $form)
     {
-        $From = $this->EmailFrom;
-        $To = $this->EmailTo;
+        $From    = $this->EmailFrom;
+        $To      = $this->EmailTo;
         $Subject = $this->EmailSubject;
-        $email = new Email($From, $To, $Subject);
-        $replyTo = $this->EnquiryFormFields()->filter(['FieldType' => 'Email'])->First();
+        $email   = new Email($From, $To, $Subject);
+
+        $replyTo = $this->EnquiryFormFields()
+            ->filter(['FieldType' => 'Email'])->First();
         if ($replyTo) {
             $post_field = $this->keyGen($replyTo->FieldName, $replyTo->SortOrder);
             if (isset($data[$post_field]) && Email::is_valid_address($data[$post_field])) {
@@ -195,7 +245,9 @@ class EnquiryPageController extends PageController
         //abuse / tracking
         $ip = EnquiryPage::get_client_ip();
         if ($ip) {
-            $email->getSwiftMessage()->getHeaders()->addTextHeader('X-Sender-IP', $ip);
+            $email->getSwiftMessage()
+                ->getHeaders()
+                ->addTextHeader('X-Sender-IP', $ip);
         }
 
         $templateData = $this->getTemplateData($data);
@@ -227,36 +279,48 @@ class EnquiryPageController extends PageController
     {
         $this->response = new HTTPResponse();
         $this->response->addHeader('Content-Type', 'image/jpeg');
-        $width = 60;
-        $height = Config::inst()->get('Axllent\EnquiryPage\EnquiryPage', 'captcha_img_height');
+        $width  = 60;
+        $height = Config::inst()
+            ->get('Axllent\EnquiryPage\EnquiryPage', 'captcha_img_height');
         $my_image = imagecreatetruecolor($width, $height);
         imagefill($my_image, 0, 0, 0xFFFFFF);
-        $purple = imageColorAllocate($my_image, 200, 0, 255);
-        $black = imageColorAllocate($my_image, 255, 255, 255);
-        $green = imageColorAllocate($my_image, 22, 255, 2);
+        $purple         = imageColorAllocate($my_image, 200, 0, 255);
+        $black          = imageColorAllocate($my_image, 255, 255, 255);
+        $green          = imageColorAllocate($my_image, 22, 255, 2);
         $random_colours = [$purple, $green, $black];
         // add noise
         for ($c = 0; $c < 150; $c++) {
-            $x = rand(0, $width-1);
-            $y = rand(0, $height-1);
-            imagesetpixel($my_image, $x, $y, $random_colours[array_rand($random_colours)]);
+            $x = rand(0, $width - 1);
+            $y = rand(0, $height - 1);
+            imagesetpixel(
+                $my_image,
+                $x,
+                $y,
+                $random_colours[array_rand($random_colours)]
+            );
         }
-        $x = rand(1, 15);
-        $token = rand(1000, 9999);
+        $x       = rand(1, 15);
+        $token   = rand(1000, 9999);
         $numbers = str_split($token);
         foreach ($numbers as $number) {
             $y = rand(1, $height - 20);
             imagestring($my_image, 5, $x, $y, $number, 0x000000);
-            $x = $x+12;
+            $x = $x + 12;
         }
         $request->getSession()->set('customcaptcha', EnquiryPage::get_hash($token));
         $this->response->setBody(imagejpeg($my_image));
         imagedestroy($my_image);
+
         return $this->response;
     }
 
+    /**
+     * Get captcha label
+     *
+     * @return string
+     */
     public function getCaptchaLabel()
     {
-        return ($this->CaptchaText) ? $this->CaptchaText : 'Verification Image';
+        return ($this->CaptchaText) ? $this->CaptchaText : 'Verification image';
     }
 }